requirements.txt
streamlit==1.40.0
pandas==2.2.3

google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.153.0
google-auth-httplib2==0.2.0
httplib2==0.22.0
